Add simple damping
- Introduces a basic damping term to reduce oscillations

+++ cpp/include/CircleWorldCollider.h
@@ -5,25 +5,25 @@
 namespace sim {
     class CircleCollider : public WorldCollider {
         public:
-            CircleCollider(Vector2 center, float radius);
+            CircleCollider(Vector2 center, double radius);
             ~CircleCollider();
             Vector2 getCenter() { return center; }
-            float getRadius() { return radius; }
+            double getRadius() { return radius; }
         protected:
             Vector2 center;
-            float radius;
+            double radius;
     };
 
     class InnerCircleCollider : public CircleCollider {
     public:
-        InnerCircleCollider(Vector2 center, float radius);
+        InnerCircleCollider(Vector2 center, double radius);
         
         bool collide(Particle* p) override;
     };
 
     class OuterCircleCollider : public CircleCollider {
     public:
-        OuterCircleCollider(Vector2 center, float radius);
+        OuterCircleCollider(Vector2 center, double radius);
         
         bool collide(Particle* p) override;
     };
