+++ cpp/include/Particle.h
@@ -4,7 +4,7 @@
 namespace sim {
     class Particle {
     public:
-        Particle(Vector2 pos, float mass = 1.0f, bool pinned = false);
+        Particle(Vector2 pos, double mass = 1.0f, bool pinned = false);
         ~Particle();
 
         // Core function
@@ -17,10 +17,10 @@ namespace sim {
         void setPosition(const Vector2& p) { position = p; }
         void setPrevPosition(const Vector2& p) { prev_position = p; }
 
-        float getRadius() const { return radius; }
-        float getMass() const { return mass; }
+        double getRadius() const { return radius; }
+        double getMass() const { return mass; }
         // inverse mass: 0 means immovable
-        float getInvMass() const { return (mass <= 0.0f) ? 0.0f : (1.0f / mass); }
+        double getInvMass() const { return (mass <= 0.0f) ? 0.0f : (1.0f / mass); }
 
         bool isPinned() const { return pinned; }
 
@@ -28,8 +28,8 @@ namespace sim {
         Vector2 position;
         Vector2 prev_position;
         Vector2 force_accum;
-        float radius;
-        float mass;
+        double radius;
+        double mass;
         bool pinned;
     };
 }
