+++ cpp/include/PlaneWorldCollider.h
@@ -5,15 +5,15 @@
 namespace sim {
     class PlaneCollider : public WorldCollider {
     public:
-        PlaneCollider(Vector2 normal, float d);
+        PlaneCollider(Vector2 normal, double d);
         ~PlaneCollider();
         
         bool collide(Particle* p) override;
         Vector2 getNormal() { return normal; }
-        float getDistance() { return d; }
+        double getDistance() { return d; }
 
     private:
         Vector2 normal;
-        float d;
+        double d;
     };
 }
