+++ cpp/src/CircleWorldCollider.cpp
@@ -2,20 +2,20 @@
 
 using namespace sim;
 
-CircleCollider::CircleCollider(Vector2 center, float radius)
+CircleCollider::CircleCollider(Vector2 center, double radius)
     : center(center), radius(radius) {}
 
 CircleCollider::~CircleCollider() {
     std::cout << "CircleCollider destroyed\n";
 }
 
-InnerCircleCollider::InnerCircleCollider(Vector2 center, float radius)
+InnerCircleCollider::InnerCircleCollider(Vector2 center, double radius)
     : CircleCollider(center, radius) {}
 
 bool InnerCircleCollider::collide(Particle* p) {
     Vector2 toP = p->getPosition() - center;
-    float dist = toP.length();
-    float maxDist = radius - p->getRadius();
+    double dist = toP.length();
+    double maxDist = radius - p->getRadius();
 
     if (dist > maxDist) {
         Vector2 n = toP / dist; // normalized
@@ -25,13 +25,13 @@ bool InnerCircleCollider::collide(Particle* p) {
     return false;
 }
 
-OuterCircleCollider::OuterCircleCollider(Vector2 center, float radius)
+OuterCircleCollider::OuterCircleCollider(Vector2 center, double radius)
     : CircleCollider(center, radius) {}
 
 bool OuterCircleCollider::collide(Particle* p) {
     Vector2 toP = p->getPosition() - center;
-    float dist = toP.length();
-    float maxDist = radius + p->getRadius();
+    double dist = toP.length();
+    double maxDist = radius + p->getRadius();
 
     if (dist < maxDist) {
         Vector2 n = toP / dist; // normalized
